move logger out of zuora module
