+++ handlers/discount-api/test/applyDiscountIntegration.test.ts
@@ -4,7 +4,7 @@
 import type { EmailMessageWithUserId } from '@modules/email/email';
 import type { Stage } from '@modules/stage';
 import { zuoraDateFormat } from '@modules/zuora/common';
-import { Logger } from '@modules/zuora/logger';
+import { Logger } from '@modules/logger';
 import { cancelSubscription } from '@modules/zuora/subscription';
 import { ZuoraClient } from '@modules/zuora/zuoraClient';
 import dayjs from 'dayjs';
