+++ handlers/product-switch-api/test/amendments.test.ts
@@ -1,5 +1,5 @@
 import type { BearerTokenProvider } from '@modules/zuora/bearerTokenProvider';
-import type { Logger } from '@modules/zuora/logger';
+import type { Logger } from '@modules/logger';
 import { ZuoraClient } from '@modules/zuora/zuoraClient';
 import { getLastAmendment } from '../src/amendments';
