+++ handlers/update-supporter-plus-amount/test/updateSupporterPlusAmountIntegration.test.ts
@@ -1,7 +1,7 @@
 import console from 'console';
 import { sendEmail } from '@modules/email/email';
 import { getProductCatalogFromApi } from '@modules/product-catalog/api';
-import { Logger } from '@modules/zuora/logger';
+import { Logger } from '@modules/logger';
 import { ZuoraClient } from '@modules/zuora/zuoraClient';
 import dayjs from 'dayjs';
 import { createThankYouEmail } from '../src/sendEmail';
