+++ handlers/update-supporter-plus-amount/test/updateSupporterPlusAmounts.test.ts
@@ -1,6 +1,6 @@
 import { generateProductCatalog } from '@modules/product-catalog/generateProductCatalog';
 import type { ProductCatalog } from '@modules/product-catalog/productCatalog';
-import { Logger } from '@modules/zuora/logger';
+import { Logger } from '@modules/logger';
 import { zuoraSubscriptionResponseSchema } from '@modules/zuora/zuoraSchemas';
 import zuoraCatalogFixture from '../../../modules/zuora-catalog/test/fixtures/catalog-code.json';
 import { getSupporterPlusData } from '../src/updateSupporterPlusAmount';
