+++ modules/zuora/test/zuoraClientIntegration.test.ts
@@ -4,7 +4,7 @@
  * @group integration
  */
 
-import { Logger } from '@modules/zuora/logger';
+import { Logger } from '@modules/logger';
 import { BearerTokenProvider } from '../src/bearerTokenProvider';
 import { getOAuthClientCredentials } from '../src/oAuthCredentials';
 import { ZuoraClient } from '../src/zuoraClient';
